Add Transformer.res_power_losses
